server/requirements.txt
ariadne>=0.22
flask-sqlalchemy>=3.1.1
alembic>=1.13.1
flask-cors>=3.0.10
flask>=3.0.2
psycopg2-binary>=2.9.9
pytz>=2024.1
pyuwsgi>=2.0.23